+++ Core/include/Acts/Seeding/SeedFilterConfig.hpp
@@ -10,6 +10,7 @@
 
 #include "Acts/Definitions/Units.hpp"
 #include "Acts/Seeding/SeedFinderUtils.hpp"
+#include "Acts/Seeding/SeedConfirmationRange.hpp"
 
 // System include(s).
 #include <cstddef>
